requirements.txt
fastapi==0.110.3
uvicorn==0.25.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
requests==2.31.0
aiohttp==3.9.5
asyncpg==0.29.0
upstash-redis==1.2.0
qstash==2.0.0
pandas==2.1.4